Zrefaktorowany kod htmla, prześlę w osobnym branchu. Majki pracuje nad driversami
