+++ clients/views/__init__.py
@@ -1,0 +1,3 @@
+from .auth_views import client_login
+from .ride_views import client_home, book_ride_endpoint, courses_rest_view
+from .registration_views import ClientRegistrationView
